+++ pkg/influx/api.go
@@ -1,17 +1,13 @@
 package influx
 
 import (
-	"context"
-	"fmt"
 	"net/http"
 	"time"
 
 	"github.com/cortexproject/cortex/pkg/cortexpb"
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
-	"github.com/grafana/influx2cortex/pkg/errorx"
 	"github.com/grafana/influx2cortex/pkg/remotewrite"
-	"github.com/pkg/errors"
 	"github.com/prometheus/client_golang/prometheus"
 	"github.com/prometheus/client_golang/prometheus/promauto"
 	"github.com/weaveworks/common/middleware"
@@ -80,51 +76,4 @@ func (a *API) handleSeriesPush(w http.ResponseWriter, r *http.Request) {
 	level.Debug(a.logger).Log("msg", "successful series write", "len", len(rwReq.Timeseries))
 
 	w.WriteHeader(http.StatusNoContent) // Needed for Telegraf, otherwise it tries to marshal JSON and considers the write a failure.
-}
-
-// handleError tries to extract an errorx.Error from the given error, logging
-// and setting the http response code as needed. All non-errorx errors are
-// considered internal errors. Please do not try to fix error categorization in
-// this function. All client errors should be categorized as an errorx at the
-// site where they are thrown.
-func (a *API) handleError(w http.ResponseWriter, r *http.Request, err error) {
-	var statusCode int
-	var httpErrString string
-	var errx errorx.Error
-	switch {
-	case errors.As(err, &errx):
-		httpErrString = errx.Message()
-		statusCode = errx.HTTPStatusCode()
-		err = errx
-	case errors.Is(err, context.DeadlineExceeded) || isGRPCTimeout(err):
-		httpErrString = "network timeout"
-		statusCode = http.StatusGatewayTimeout
-	case errors.Is(err, context.Canceled):
-		// Note: It seems unlikely this can happen other than as a timeout, so we
-		// should call it an internal error.
-		httpErrString = "request cancelled"
-		statusCode = http.StatusInternalServerError
-	case isNetworkTimeout(err):
-		if r.Body != nil {
-			// Try to read 1 byte from the request body. If it fails with the same error
-			// it means the timeout occurred while reading the request body, so it's a 408.
-			if _, readErr := r.Body.Read([]byte{0}); isNetworkTimeout(readErr) {
-				httpErrString = "response timeout"
-				statusCode = http.StatusRequestTimeout
-				break
-			}
-			httpErrString = "network timeout"
-			statusCode = http.StatusGatewayTimeout
-		}
-	default:
-		httpErrString = "uncategorized error"
-		statusCode = http.StatusInternalServerError
-	}
-	if statusCode < 500 {
-		level.Info(a.logger).Log("msg", httpErrString, "response_code", statusCode, "err", tryUnwrap(err))
-	} else if statusCode >= 500 {
-		level.Warn(a.logger).Log("msg", httpErrString, "response_code", statusCode, "err", tryUnwrap(err))
-	}
-	a.recorder.measureProxyErrors(fmt.Sprintf("%T", err))
-	http.Error(w, httpErrString, statusCode)
 }
